% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blineShift.R
\name{blineShift}
\alias{blineShift}
\title{Shift the baseline to 0}
\usage{
blineShift(resp, logc, wndw)
}
\arguments{
\item{resp}{Numeric, the response values}

\item{logc}{Numeric, the log10 concentration values}

\item{wndw}{Numeric, the threshold window}
}
\value{
A numeric vector containing the shifted response values
}
\description{
\code{blineShift} Takes in dose-response data and shifts the baseline
to 0 based on the window.
}
\note{
This function is not exported and is not inteded to be used by the user.
}
\seealso{
\code{\link{mc3_mthds}}, \code{\link{mc3}}
}

