% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAssayQ.R
\name{.buildAssayQ}
\alias{.buildAssayQ}
\title{Generate query for assay information}
\usage{
.buildAssayQ(out, tblo, fld = NULL, val = NULL, add.fld = NULL)
}
\arguments{
\item{out}{Character, the default fields to include}

\item{tblo}{Integer, the order to send the fields to prepOutput}

\item{fld}{Character, the field(s) to query/subset on}

\item{val}{List, vectors of values for each field to query/subset on. Must 
be in the same order as 'fld'.}

\item{add.fld}{Character, additional field(s) to include, but not query/
subset on}
}
\value{
A character containing the query to send to tcplQuery
}
\description{
\code{.buildAssayQ} generates a query string to load assay information
}
