% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemQueries.R
\name{.ChemListQ}
\alias{.ChemListQ}
\title{Function to support queries by chemical}
\usage{
.ChemListQ(field, val)
}
\arguments{
\item{field}{Character, the field to query on}

\item{val}{Vector of values to subset on}
}
\description{
\code{.ChemListQ} creates tcplLoadChemList query string
}
\note{
This function is not exported and not intended to be used by the user.
}
\seealso{
\code{\link{tcplLoadChemList}}
}
