% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemQueries.R
\name{.ChemQ}
\alias{.ChemQ}
\title{Function to support queries by chemical}
\usage{
.ChemQ(field, val, exact)
}
\arguments{
\item{field}{Character, the field to query on}

\item{val}{Vector of values to subset on}

\item{exact}{Logical, should chemical names be considered exact?}
}
\description{
\code{.ChemQ} creates tcplLoadChem query string
}
\note{
This function is not exported and not intended to be used by the user.
}
\seealso{
\code{\link{tcplLoadData}}

\code{\link{tcplLoadChem}}
}
