% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNames.R
\name{.convertNames}
\alias{.convertNames}
\title{Convert assay names to their abbreviations}
\usage{
.convertNames(names)
}
\arguments{
\item{names}{Character, strings to convert}
}
\value{
The same character vector given with any name strings converted to 
the abbreviated version
}
\description{
\code{.convertNames} converts the assay names as they appear in the tcpl
database to their respective abbreviations
}
