% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateHeat.R
\name{.plateHeat}
\alias{.plateHeat}
\title{Plot plate heatmap}
\usage{
.plateHeat(vals, rowi, coli, wllt, wllq, rown, coln, main, arng)
}
\arguments{
\item{vals}{Numeric, the well values}

\item{rowi}{Integer, the row index}

\item{coli}{Integer, the column index}

\item{wllt}{Character, the well type}

\item{wllq}{Logical, the well quality}

\item{rown}{Integer, the number of rows on the plate}

\item{coln}{Integer, the number of columns on the plate}

\item{main}{Character of length 1, the title/main}

\item{arng}{Numeric of length 2, the minimum and maximum values to constrain the color scale}
}
\description{
Plot plate heatmap, to be used with tcplPlotPlate
}
\note{
Optimized for an output with height = 20/3, width = 10, and 
pointsize = 10
}
