% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepField.R
\name{.prepField}
\alias{.prepField}
\title{Paste appropriate table name to field name}
\usage{
.prepField(fld, tbl, db)
}
\arguments{
\item{fld}{Character, the table fields}

\item{tbl}{Character, the possible tables}

\item{db}{Character, the database containing the tables}
}
\description{
Paste appropriate table name to field name
}
\details{
The function loops through the given tables, and for each field i it assigns
the last table containing i to i. ORDER OF FLD MATTERS!!
}
