% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplot2Utils.R
\name{dynamic_table_trunc}
\alias{dynamic_table_trunc}
\title{dynamic_table_trunc
Dynamically truncate lengths of column values of they are longer than a
calculated width. Strings contained in verbose table output can be very long, 
and this function ensures that the 128 character row limit is kept while
growing/shrinking default columns widths that may or may not need the space.}
\usage{
dynamic_table_trunc(tbl = NULL, all_cols)
}
\arguments{
\item{tbl}{data.table with potential long values to truncate}

\item{all_cols}{every annotation column to assign lengths}
}
\value{
altered data table with truncated strings
}
\description{
dynamic_table_trunc
Dynamically truncate lengths of column values of they are longer than a
calculated width. Strings contained in verbose table output can be very long, 
and this function ensures that the 128 character row limit is kept while
growing/shrinking default columns widths that may or may not need the space.
}
