% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplot2Utils.R
\name{get_plot_caption}
\alias{get_plot_caption}
\title{get_plot_caption
Generate plot caption given number of flags and length of unique elements}
\usage{
get_plot_caption(dat = NULL)
}
\arguments{
\item{dat}{data table including all required flags}
}
\value{
a string caption for the individual plot
}
\description{
get_plot_caption
Generate plot caption given number of flags and length of unique elements
}
