% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplot2Utils.R
\name{get_plot_title}
\alias{get_plot_title}
\title{get_plot_title
Generate plot title given number of rows and length of unique elements}
\usage{
get_plot_title(dat = NULL, type = "mc", compare = "m4id", verbose = FALSE)
}
\arguments{
\item{dat}{data table with all required conc/resp data; each row will extend comparison}

\item{type}{string of mc or sc indicating if it is single or multi conc}

\item{compare}{Character vector, the field(s) samples were joined on for comparison}

\item{verbose}{bool if hitc should be included in the title}
}
\value{
a string title for the individual plot
}
\description{
get_plot_title
Generate plot title given number of rows and length of unique elements
}
