% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplot2Utils.R
\name{get_verbose_tables}
\alias{get_verbose_tables}
\title{get_verbose_tables
Generates 'gt' package tables containing annotations and level 5 metrics}
\usage{
get_verbose_tables(dat = NULL, type = "mc", compare = "m4id", flags = FALSE)
}
\arguments{
\item{dat}{data table with all level 5 metrics and mapped annotations}

\item{type}{string of mc or sc indicating if it is single or multi conc}

\item{compare}{Character vector, the field(s) samples were joined on for comparison}

\item{flags}{bool if flags should be included in the table footer}
}
\value{
list of 2 GT tables
}
\description{
get_verbose_tables
Generates 'gt' package tables containing annotations and level 5 metrics
}
