% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlaceFunc.R
\name{interlaceFunc}
\alias{interlaceFunc}
\title{Calculate the weighted mean of a square to detect interlace effect}
\usage{
interlaceFunc(val, intq, coli, rowi, apid, r)
}
\arguments{
\item{val}{Numeric, the well values}

\item{intq}{Numeric, interlace quadrant}

\item{coli}{Integer, the well column index}

\item{rowi}{Integer, the well row index}

\item{apid}{Character, the assay plate id}

\item{r}{Integer, the number of wells from the center well (in one 
direction) to make the square}
}
\description{
\code{interlaceFunc} calculates the distance weighted mean of square regions
from a 384-well plate that is interlaced onto a 1536 well plate to detect
non-random signals coming from the source plate
}
\seealso{
\code{\link{MC6_Methods}}, \code{\link{Method functions}}, 
\code{\link{mc6}}
}
