% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invitrodb_dd}
\alias{invitrodb_dd}
\title{Short descriptions of fields for different tables are stored in a data dictionary.}
\format{
A data frame with 44 rows and 3 variables:
\describe{
 \item{invitrodb_table}{Table of the data dictionary}
  \item{invitrodb_field}{Field of the data dictionary}
   \item{description}{Description}
}
}
\source{
ToxCast database
}
\usage{
invitrodb_dd
}
\description{
Short descriptions of fields for different tables are stored in a data dictionary.
}
\keyword{datasets}
