% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.odd.R
\name{is.odd}
\alias{is.odd}
\title{Check for odd numbers}
\usage{
is.odd(x)
}
\arguments{
\item{x}{An integer}
}
\value{
\code{TRUE} for odd integers and \code{FALSE} for even integers.
}
\description{
\code{is.odd} takes an integer vector, \code{x}, and returns TRUE for odd
integers.
}
\seealso{
Other tcpl abbreviations: \code{\link{lu}},
  \code{\link{lw}}, \code{\link{sink.reset}}
}

