% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw.R, R/tcplFit.R
\name{lw}
\alias{lw}
\title{Abbreviation for \code{length(which(x))}}
\usage{
lw(x)

lw(x)
}
\arguments{
\item{x}{A logical}
}
\value{
The length of the \code{TRUE} values in \code{x}

The length of the \code{TRUE} values in \code{x}
}
\description{
\code{lw} takes a logical vector, \code{x}, and returns 
\code{length(which(x))}.

\code{lw} takes a logical vector, \code{x}, and returns
\code{length(which(x))}.
}
\seealso{
\code{\link{length}}, \code{\link{which}}

\code{\link{length}}, \code{\link{which}}

Other tcpl abbreviations: 
\code{\link{is.odd}()},
\code{\link{lu}()},
\code{\link{sink.reset}()}

Other tcpl abbreviations: 
\code{\link{is.odd}()},
\code{\link{lu}()},
\code{\link{sink.reset}()}
}
\concept{tcpl abbreviations}
