% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc6_mthds.R
\name{MC6_Methods}
\alias{MC6_Methods}
\alias{mc6_mthds}
\title{Load list of level 6 multiple-concentration flag methods}
\usage{
mc6_mthds()
}
\value{
A list functions
}
\description{
\code{mc6_mthds} returns a list of flag methods to be used 
during level 6 multiple-concentration processing.
}
\section{Available Methods}{


More information about the level 6 multiple-concentration processing is 
available in the package vignette, "Pipeline_Overview."

\describe{
  \item{modl.directionality.fail}{The modl.directionality.fail flag 
  identifies concentration series where, had the model direction been 
  opposite, more responses would have exceeded the cutoff, i.e., series 
  where the number of responses in absolute value which are greater than the 
  absolute value of the cutoff exceeds twice the number of responses which 
  are more extreme than the cutoff (< cutoff when cutoff is negative, 
  > cutoff when cutoff is positive).}
  \item{low.nrep}{The low.nrep flag identifies concentration series where 
  the average number of replicates per concentration tested is less than 2.}
  \item{low.nconc}{The low.nconc flag identifies concentration series where 
  the number of concentrations tested is greater or equal to 4.}
  \item{bmd.highconc}{The bmd.highconc flag identifies concentration series 
  with a steep curve, i.e., series where the bmd is greater than 50% of the 
  range of concentrations tested.}
  \item{bmd.lowconc}{The bmd.lowconc flag identifies concentration series 
  with a shallow curve, i.e., series where the bmd is less than 50% of the 
  range of concentrations tested.}
  \item{bmd.high}{The bmd.high flag attempts to identify noisy concentration
  series by flagging series where the bmd is greater than the ac50.}
  \item{singlept.hit.high}{The singlept.hit.high flag identifies 
  concentration series where the median response was greater than 3*bmad 
  only at the highest tested concentration and the series had an active 
  hit-call.}
  \item{singlept.hit.mid}{The singlept.hit.mid flag identifies concentration 
  series where the median response was greater than 3*bmad at only one 
  concentration (not the highest tested concentration) and the series had 
  an active hit-call.}
  \item{multipoint.neg}{The multipoint.neg flag identifies concentration 
  series with response medians greater than 3*bmad at multiple 
  concentrations and an inactive hit-call.}
  \item{gnls.lowconc}{The gnls.lowconc flag identifies concentration series
  where the gain-loss model won, the gain AC50 is less than the minimum 
  tested concentration, and the loss AC50 is less than the mean tested 
  concentration.}
  \item{noise}{The noise flag attempts to identify noisy concentration
  series by flagging series where the root mean square error for the series
  is greater than the cutoff for the assay endpoint.}
  \item{border}{The border flag identifies concentration 
  series which have borderline activity, i.e., series where the top 
  parameter of the winning model was greater than or equal to 0.8*cutoff 
  and less than or equal to 1.2*cutoff.}
  \item{overfit.hit}{The overfit.hit flag recalculates the model winner 
  after applying a small sample correction factor to the AIC values. If the 
  hit-call would be changed after applying the small sample correction 
  factor the series is flagged. Series with less than 5 concentrations where
  the hill model won and series with less than 7 concentrations where the 
  gain-loss model won are automatically flagged.}
  \item{efficacy.50}{The efficacy.50 flag identifies concentration series 
  with efficacy values (either the modeled top parameter for the winning
  model or the maximum median response) less than 50 for percent activity
  data or log2(1.5) for fold induction data}
  \item{ac50.lowconc}{The ac50.lowconc flag identifies concentration series 
  with ac50 values less than the minimum tested concentration.}
  \item{viability.gnls}{The viability.gnls flag identifies concentration series 
  of cell viability assays which were fit with gain-loss as the winning 
  model and the series had an active hit-call.}
}
}

\seealso{
\code{\link{mc6}}, \code{\link{Method functions}} to query what
methods get applied to each aeid
}
