% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_test}
\alias{mc_test}
\title{List of lists containing queries sent to tcplQuery associated with each test
case. Each list also contains the associated ids with each case. Only meant
to be used with automated testing with mocking for mc data.}
\format{
A list with 30 items:
\describe{
  \item{tcplConfQuery}{Data table with 1 row and 2 columns used for each test
  case for establishing connection using tcplConf. This data table mocks the
  response one would get from connecting with invitrodb.}
  \item{mc0_by_m0id}{List containing the queries used for loading mc0 data by
  m0id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'm0id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc0_by_acid}{List containing the queries used for loading mc0 data by
  acid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'acid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc1_by_m1id}{List containing the queries used for loading mc1 data by
  m1id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'm1id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc1_by_acid}{List containing the queries used for loading mc1 data by
  acid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'acid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc2_by_m2id}{List containing the queries used for loading mc2 data by
  m2id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'm2id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc2_by_acid}{List containing the queries used for loading mc2 data by
  acid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'acid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc3_by_m3id}{List containing the queries used for loading mc3 data by
  m3id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'm3id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc3_by_aeid}{List containing the queries used for loading mc3 data by
  aeid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'aeid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc4_by_m4id}{List containing the queries used for loading mc4 data by
  m4id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'm4id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc4_by_aeid}{List containing the queries used for loading mc4 data by
  aeid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'aeid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc5_by_m5id}{List containing the queries used for loading mc5 data by
  m5id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'm5id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc5_by_aeid}{List containing the queries used for loading mc5 data by
  aeid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'aeid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc6_by_m6id}{List containing the queries used for loading mc6 data by
  m6id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'm6id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc6_by_aeid}{List containing the queries used for loading mc6 data by
  aeid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'aeid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc7_by_m7id}{List containing the queries used for loading mc7 data by
  m7id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'm7id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mc7_by_aeid}{List containing the queries used for loading mc7 data by
  aeid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'aeid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{mcagg_by_aeid}{List containing the queries used for loading mc 'agg'
  data by aeid via tcplLoadData. Each query has an associated data table 
  response for mocking an actual connection. Contains one 'aeid' labeled item
  storing the id used to load the data, for use in tests.}
  \item{plot_single_m4id}{List containing the queries used for loading and
  plotting mc data by m4id via tcplPlot. Each query has an associated data
  table response for mocking an actual connection. Contains one 'm4id'
  labeled item storing the id used to load the data, for use in tests.}
  \item{plot_multiple_m4id}{List containing the queries used for loading and
  plotting mc data by multiple m4ids via tcplPlot. Each query has an
  associated data table response for mocking an actual connection. Contains
  one 'm4id' labeled item storing the ids used to load the data, for use in
  tests.}
  \item{plot_single_aeid}{List containing the queries used for loading and
  plotting mc data by aeid via tcplPlot. Each query has an associated data
  table response for mocking an actual connection. Contains one 'aeid'
  labeled item storing the id used to load the data, for use in tests.}
  \item{plot_multiple_aeid}{List containing the queries used for loading and
  plotting mc data by multiple aeids via tcplPlot. Each query has an
  associated data table response for mocking an actual connection. Contains
  one 'aeid' labeled item storing the ids used to load the data, for use in
  tests.}
  \item{plot_single_spid}{List containing the queries used for loading and
  plotting mc data by spid/aeid via tcplPlot. Each query has an associated
  data table response for mocking an actual connection. Contains 'spid' and
  'aeid' labeled items storing the ids used to load the data, for use in
  tests.}
  \item{plot_multiple_spid}{List containing the queries used for loading and
  plotting mc data by multiple spids/aeid via tcplPlot. Each query has an
  associated data table response for mocking an actual connection. Contains
  'spid' and 'aeid' labeled items storing the ids used to load the data, for
  use in tests.}
  \item{plot_single_m4id_compare}{List containing the queries used for 
  loading and plotting compared mc data by m4id via tcplPlot. Each query has
  an associated data table response for mocking an actual connection.
  Contains 'm4id' and 'compare.m4id' labeled items storing the ids used to
  load the data, for use in tests.}
  \item{plot_multiple_m4id_compare}{List containing the queries used for loading and
  plotting compared mc data by multiple m4ids via tcplPlot. Each query has an
  associated data table response for mocking an actual connection. Contains 
  'm4id' and 'compare.m4id' labeled items storing the ids used to load the 
  data, for use in tests.}
  \item{plot_single_aeid_compare}{List containing the queries used for
  loading and plotting compared mc data by aeid via tcplPlot. Each query has
  an associated data table response for mocking an actual connection.
  Contains 'aeid' and 'compare.aeid' labeled items storing the ids used to
  load the data, for use in tests.}
  \item{plot_multiple_aeid_compare}{List containing the queries used for
  loading and plotting compared mc data by multiple aeids via tcplPlot. Each
  query has an associated data table response for mocking an actual
  connection. Contains 'aeid' and 'compare.aeid' labeled items storing the
  ids used to load the data, for use in tests.}
  \item{plot_single_spid_compare}{List containing the queries used for
  loading and plotting compared mc data by spid/aeid via tcplPlot. Each query
  has an associated data table response for mocking an actual connection.
  Contains 'spid', 'compare.spid', 'aeid', and 'compare.aeid' labeled items
  storing the ids used to load the data, for use in tests.}
  \item{plot_multiple_spid_compare}{List containing the queries used for
  loading and plotting compared mc data by multiple spids/aeid via tcplPlot.
  Each query has an associated data table response for mocking an actual
  connection. Contains 'spid', 'compare.spid', 'aeid', and 'compare.aeid'
  labeled items storing the ids used to load the data, for use in tests.}
}
}
\source{
ToxCast database
}
\usage{
mc_test
}
\description{
List of lists containing queries sent to tcplQuery associated with each test
case. Each list also contains the associated ids with each case. Only meant
to be used with automated testing with mocking for mc data.
}
\keyword{datasets}
