% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_vignette}
\alias{mc_vignette}
\title{List with multi-concentration data for the vignette}
\format{

}
\usage{
mc_vignette
}
\description{
This dataset is a list with 6 data.tables (mc0,mc1,mc2,mc3,mc4,mc5).
}
\keyword{datasets}
