% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcdat}
\alias{mcdat}
\title{A subset of ToxCast data showing changes in the activity of the intracellular estrogen receptor.}
\format{
A data frame with 14183 rows and 10 variables:
\describe{
 \item{spid}{sample ID}
  \item{apid}{assay plate ID}
   \item{rowi}{well-plate row number}
    \item{coli}{well-plate column number}
     \item{wllt}{well type}
      \item{wllq}{well quality}
       \item{conc}{concentration in micromolar}
        \item{rval}{raw assay component readout value}
         \item{srcf}{source file containing the data}
          \item{acsn}{assay component source name}
}
}
\source{
Toxcast database
}
\usage{
mcdat
}
\description{
The example dataset is used to illustrate how the user can pipeline multiple-concentration  
data from chemical screening using tcplLite.
}
\keyword{datasets}
