% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplot2Utils.R
\name{MORELETTERS}
\alias{MORELETTERS}
\title{MORELETTERS
Extends LETTERS recursively to any number of letters}
\usage{
MORELETTERS(i)
}
\arguments{
\item{i}{numeric vector of all needed letter combinations, i.e. 1:nrow(dat)}
}
\value{
char vector containing all letter/combinations
}
\description{
MORELETTERS
Extends LETTERS recursively to any number of letters
}
\author{
https://stackoverflow.com/a/25881167
}
