% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerMthd.R
\name{registerMthd}
\alias{registerMthd}
\title{Add a new analysis method}
\usage{
registerMthd(lvl, mthd, desc, nddr = 0L, type)
}
\arguments{
\item{lvl}{Integer of length 1, the level for the analysis method}

\item{mthd}{Character, the name of the method}

\item{desc}{Character, same length as mthd, the method description}

\item{nddr}{Integer, 0 or 1, 1 if the method requires loading the dose-
response data}

\item{type}{Character of length 1, the data type, "sc" or "mc"}
}
\description{
\code{registerMthd} registers a new analysis method to the tcpl 
databases.
}
\details{
'mthd' must match a corresponding function name in the functions that load
the methods, ie. \code{mc2_mthds}. 'nddr' only applies to level 6 
methods.
}
