% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplot2Utils.R
\name{round_n}
\alias{round_n}
\title{round_n
General function to round/shorten values for plotting tables}
\usage{
round_n(x, n = 3)
}
\arguments{
\item{x}{numeric value}

\item{n}{numeric number of decimal places to round to}
}
\value{
format string in decimal or scientific notation
}
\description{
round_n
General function to round/shorten values for plotting tables
}
