% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_test}
\alias{sc_test}
\title{List of lists containing queries sent to tcplQuery associated with each test
case. Each list also contains the associated ids with each case. Only meant
to be used with automated testing with mocking for sc data.}
\format{
A list with 20 items:
\describe{
  \item{tcplConfQuery}{Data table with 1 row and 2 columns used for each test
  case for establishing connection using tcplConf. This data table mocks the
  response one would get from connecting with invitrodb.}
  \item{sc0_by_s0id}{List containing the queries used for loading sc0 data by
  s0id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 's0id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{sc0_by_acid}{List containing the queries used for loading sc0 data by
  acid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'acid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{sc1_by_s1id}{List containing the queries used for loading sc1 data by
  s1id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 's1id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{sc1_by_acid}{List containing the queries used for loading sc1 data by
  acid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'acid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{sc2_by_s2id}{List containing the queries used for loading sc2 data by
  s2id via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 's2id' labeled item storing the
  id used to load the data, for use in tests.}
  \item{sc2_by_aeid}{List containing the queries used for loading sc2 data by
  aeid via tcplLoadData. Each query has an associated data table response for
  mocking an actual connection. Contains one 'aeid' labeled item storing the
  id used to load the data, for use in tests.}
  \item{scagg_by_aeid}{List containing the queries used for loading sc 'agg'
  data by aeid via tcplLoadData. Each query has an associated data table 
  response for mocking an actual connection. Contains one 'aeid' labeled item
  storing the id used to load the data, for use in tests.}
  \item{plot_single_s2id}{List containing the queries used for loading and
  plotting sc data by s2id via tcplPlot. Each query has an associated data
  table response for mocking an actual connection. Contains one 's2id'
  labeled item storing the id used to load the data, for use in tests.}
  \item{plot_multiple_s2id}{List containing the queries used for loading and
  plotting sc data by multiple s2ids via tcplPlot. Each query has an
  associated data table response for mocking an actual connection. Contains
  one 's2id' labeled item storing the ids used to load the data, for use in
  tests.}
  \item{plot_single_aeid}{List containing the queries used for loading and
  plotting sc data by aeid via tcplPlot. Each query has an associated data
  table response for mocking an actual connection. Contains one 'aeid'
  labeled item storing the id used to load the data, for use in tests.}
  \item{plot_multiple_aeid}{List containing the queries used for loading and
  plotting sc data by multiple aeids via tcplPlot. Each query has an
  associated data table response for mocking an actual connection. Contains
  one 'aeid' labeled item storing the ids used to load the data, for use in
  tests.}
  \item{plot_single_spid}{List containing the queries used for loading and
  plotting sc data by spid/aeid via tcplPlot. Each query has an associated
  data table response for mocking an actual connection. Contains 'spid' and
  'aeid' labeled items storing the ids used to load the data, for use in
  tests.}
  \item{plot_multiple_spid}{List containing the queries used for loading and
  plotting sc data by multiple spids/aeid via tcplPlot. Each query has an
  associated data table response for mocking an actual connection. Contains
  'spid' and 'aeid' labeled items storing the ids used to load the data, for
  use in tests.}
  \item{plot_single_s2id_compare}{List containing the queries used for 
  loading and plotting compared sc data by s2id via tcplPlot. Each query has
  an associated data table response for mocking an actual connection.
  Contains 's2id' and 'compare.s2id' labeled items storing the ids used to
  load the data, for use in tests.}
  \item{plot_multiple_s2id_compare}{List containing the queries used for loading and
  plotting compared sc data by multiple s2ids via tcplPlot. Each query has an
  associated data table response for mocking an actual connection. Contains 
  's2id' and 'compare.s2id' labeled items storing the ids used to load the 
  data, for use in tests.}
  \item{plot_single_aeid_compare}{List containing the queries used for
  loading and plotting compared sc data by aeid via tcplPlot. Each query has
  an associated data table response for mocking an actual connection.
  Contains 'aeid' and 'compare.aeid' labeled items storing the ids used to
  load the data, for use in tests.}
  \item{plot_multiple_aeid_compare}{List containing the queries used for
  loading and plotting compared sc data by multiple aeids via tcplPlot. Each
  query has an associated data table response for mocking an actual
  connection. Contains 'aeid' and 'compare.aeid' labeled items storing the
  ids used to load the data, for use in tests.}
  \item{plot_single_spid_compare}{List containing the queries used for
  loading and plotting compared sc data by spid/aeid via tcplPlot. Each query
  has an associated data table response for mocking an actual connection.
  Contains 'spid', 'compare.spid', 'aeid', and 'compare.aeid' labeled items
  storing the ids used to load the data, for use in tests.}
  \item{plot_multiple_spid_compare}{List containing the queries used for
  loading and plotting compared sc data by multiple spids/aeid via tcplPlot.
  Each query has an associated data table response for mocking an actual
  connection. Contains 'spid', 'compare.spid', 'aeid', and 'compare.aeid'
  labeled items storing the ids used to load the data, for use in tests.}
}
}
\source{
ToxCast database
}
\usage{
sc_test
}
\description{
List of lists containing queries sent to tcplQuery associated with each test
case. Each list also contains the associated ids with each case. Only meant
to be used with automated testing with mocking for sc data.
}
\keyword{datasets}
