% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scdat}
\alias{scdat}
\title{A subset of ToxCast data showing changes in transcription factor activity for multiple targets.}
\format{
A data frame with 320 rows and 10 variables:
\describe{

 \item{spid}{sample ID}
  \item{apid}{assay plate ID}
   \item{rowi}{well-plate row number (N/A)}
    \item{coli}{well-plate column number (N/A)}
     \item{wllt}{well type (N/A)}
      \item{wllq}{well quality (N/A)}
       \item{conc}{concentration in micromolar}
        \item{rval}{raw assay component readout value}
         \item{srcf}{source file containing the data}
          \item{acsn}{assay component source name}
}
}
\source{
Toxcast database
}
\usage{
scdat
}
\description{
The example dataset is used to illustrate how the user can pipeline single-concentration  
data from chemical screening using tcplLite.
}
\keyword{datasets}
