% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.reset.R
\name{sink.reset}
\alias{sink.reset}
\title{Reset all sinks}
\usage{
sink.reset()
}
\description{
\code{sink.reset} resets all sinks and returns all output to the console.
}
\details{
\code{sink.reset} identifies all sinks with \code{sink.number} then returns
all output and messages back to the console.
}
\seealso{
\code{\link{sink}}, \code{\link{sink.number}}

Other tcpl abbreviations: 
\code{\link{is.odd}()},
\code{\link{lu}()},
\code{\link{lw}()}
}
\concept{tcpl abbreviations}
