% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplCascade.R
\name{tcplCascade}
\alias{tcplCascade}
\title{Do a cascading delete on tcpl screening data}
\usage{
tcplCascade(lvl, type, id)
}
\arguments{
\item{lvl}{Integer of length 1, the first level to delete from}

\item{type}{Character of length 1, the data type, "sc" or "mc"}

\item{id}{Integer, the id(s) to delete. See details for more information.}
}
\description{
\code{tcplCascade} deletes the data for the given id(s) starting at 
the processing level given. The delete will cascade through all subsequent 
tables.
}
\details{
The data type can be either 'mc' for mutliple concentration data, or 'sc'
for single concentration data. Multiple concentration data will be loaded
into the level tables, whereas the single concentration will be loaded into
the single tables. 

If lvl is less than 3, id is interpreted as acid(s) and if lvl is greater 
than or equal to 3, id is interpreted as aeid(s).
}
\note{
This function is not exported and not intended to be used by the user.
}
