% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpldbStats.R
\name{tcpldbStats}
\alias{tcpldbStats}
\title{Get summary statisitcs for the database}
\usage{
tcpldbStats(type = "all", val = NULL)
}
\arguments{
\item{type}{String either "all", "aeid" or "spid"}

\item{val}{integer if type = "aeid" , string if type  = "spid"}
}
\description{
\code{tcpldbStats} takes a string(type) and an optional parameter(val) to return the summary statistics on the entire tcplLite database
 When type = "all" the val is ignored. the function returns the number of distinct spid and aeids in the database at each level
 When type = "aeid", the val parameter has to be a valid aeid in the database. The function returns a 
 table consisting of the number of distinct spids at each level of processing for the aeid given in 'val'
 When type = "spid", the val parameter has to be a valid spid in the database. The function returns a 
 table consisting of the number of distinct aeids at each level of processing for the given spid in 'val'
}
