% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplDelete.R
\name{tcplDelete}
\alias{tcplDelete}
\title{Delete rows from tcpl databases}
\usage{
tcplDelete(tbl, fld, val, db)
}
\arguments{
\item{tbl}{Character, length 1, the table to delete from}

\item{fld}{Character, the field(s) to query on}

\item{val}{List, vectors of values for each field to query on. Must be in 
the same order as 'fld'.}

\item{db}{Character, the database containing the table}
}
\description{
\code{tcplDelete} deletes rows from the given table and database.
}
\note{
This function is not exported and not intended to be used by the user.
}
\seealso{
\code{\link{tcplSendQuery}}
}
