% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplot2Utils.R
\name{tcplfit2_fun}
\alias{tcplfit2_fun}
\title{tcplfit2_fun
Returns tcplfit2 function for drawing models given input}
\usage{
tcplfit2_fun(row, model, x)
}
\arguments{
\item{row}{data table with all required conc/resp data}

\item{model}{string of model to draw from tcplfit2}

\item{x}{int vector of x values to calculate curve (y) values for}
}
\value{
a tcplfit2 function call
}
\description{
tcplfit2_fun
Returns tcplfit2 function for drawing models given input
}
