% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplFit2.R
\name{tcplFit2_nest}
\alias{tcplFit2_nest}
\title{Nest dataframe into a list that is readable by tcplfit2}
\usage{
tcplFit2_nest(test)
}
\arguments{
\item{test}{}
}
\value{
a list of fitting parameters that can be consumed by tcplfit2
}
\description{
Nest dataframe into a list that is readable by tcplfit2
}
