% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplFit2.R
\name{tcplFit2_unnest}
\alias{tcplFit2_unnest}
\title{Unnest tcplfit2 parameters into a dataframe}
\usage{
tcplFit2_unnest(output)
}
\arguments{
\item{output}{}
}
\value{
list of parameters unnested and compiled into a dataframe
}
\description{
Unnest tcplfit2 parameters into a dataframe
}
