% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplGetAeid.R
\name{tcplGetAeid}
\alias{tcplGetAeid}
\title{get Aeid for endpoint name}
\usage{
tcplGetAeid(name)
}
\arguments{
\item{name}{A string that will be matched to the assay component endpoint name}
}
\description{
\code{tcplGetAeid} takes a string(name) and finds the assay component endpoint names that match the string and the 
aeids associated with those names.The function performes a regular expression like matching for strings in the 
assay component endpoint name column in the assay component endpoint table.
}
\examples{
## Store the current config settings, so they can be reloaded at the end 
## of the examples 
conf_store <- tcplConfList()
tcplConfExample()

## Search for aenm (assay name) case insensitive
tcplGetAeid("TOX21")
tcplGetAeid("tox21")

## Reset configuration
options(conf_store)

}
