% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlot.R
\name{tcplggplot}
\alias{tcplggplot}
\title{tcplggplot}
\usage{
tcplggplot(dat, lvl = 5, verbose = FALSE)
}
\arguments{
\item{dat}{data table with all required conc/resp data}

\item{lvl}{integer level of data that should be plotted
level 4 - all fit models
level 5 - all fit models and winning model with hitcall
level 6 - include all flags}

\item{verbose}{boolean should plotting include table of values next to the plot}
}
\value{

}
\description{
tcplggplot
}
