% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplggplot2.R
\name{tcplggplot2}
\alias{tcplggplot2}
\title{tcplggplot2}
\usage{
tcplggplot2(
  dat,
  type = "mc",
  compare = "m4id",
  verbose = FALSE,
  flags = FALSE,
  yrange = c(NA, NA),
  group.fld = NULL,
  group.threshold = 9
)
}
\arguments{
\item{dat}{data table with all required conc/resp data; each row will extend comparison}

\item{type}{string type of data that should be plotted
'sc' plotting - max medians plotted with hitcall
'mc' plotting - all fit models and winning model with hitcall}

\item{compare}{Character vector, the field(s) samples were joined on for comparison}

\item{verbose}{boolean should plotting include table of values next to the plot}

\item{flags}{boolean should plotting include level 6 flags in plot caption}

\item{yrange}{Integer of length 2, for overriding the y-axis range, c(<min>,<max>). 
By default, c(NA,NA).}

\item{group.fld}{string column name to group curves by when number of curves 
exceeds group.threshold, default being 'modl' for MC and 'hitc' for SC}

\item{group.threshold}{integer of length 1, number of curves where plot style
should change to instead group by a given group.fld, default of 9 -- greater 
than 8 curves}
}
\value{
A ggplot object or grob with accompanied table depending on verbose option
}
\description{
tcplggplot2
}
