% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlot.R
\name{tcplggplotCompare}
\alias{tcplggplotCompare}
\title{tcplggplotCompare}
\usage{
tcplggplotCompare(
  dat,
  compare.dat,
  lvl = 5,
  verbose = FALSE,
  flags = FALSE,
  yrange = c(NA, NA)
)
}
\arguments{
\item{dat}{data table with all required conc/resp data}

\item{compare.dat}{data table with all required conc/resp data for comparison
overlay}

\item{lvl}{integer level of data that should be plotted
level 2 - for 'sc' plotting
level 5 - for 'mc' plotting, all fit models and winning model with hitcall}

\item{verbose}{boolean should plotting include table of values next to the plot}

\item{flags}{boolean should plotting include level 6 flags in plot caption}

\item{yrange}{Integer of length 2, for overriding the y-axis range, c(<min>,<max>). 
By default, c(NA,NA).}
}
\value{
A ggplot object or grob with accompanied table depending on verbose option
}
\description{
tcplggplotCompare
}
