% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplFit2.R
\name{tcplHit2}
\alias{tcplHit2}
\title{Hitcalling with tcplfit2}
\usage{
tcplHit2(mc4, coff)
}
\arguments{
\item{mc4}{data.table with level 4 data}

\item{coff}{cutoff value for hitcalling}
}
\value{
Data.table with key value pairs of hitcalling parameters
}
\description{
Hitcalling with tcplfit2
}
