% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplListFlds.R
\name{tcplListFlds}
\alias{tcplListFlds}
\title{Load the field names for a table}
\usage{
tcplListFlds(tbl, db = getOption("TCPL_DB"))
}
\arguments{
\item{tbl}{Character of length 1, the tcpl database table}

\item{db}{Character of length 1, the tcpl database}
}
\value{
A string of field names for the given table.
}
\description{
\code{tcplListFlds} loads the column names for the given table and database.
}
\details{
This function can be particularly useful in defining the 'fld' param in the
tcplLoad- functions.
}
\examples{
## Gives the fields in the mc1 table
tcplListFlds("mc1")

}

