% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplLoadClib.R
\name{tcplLoadClib}
\alias{tcplLoadClib}
\title{Load chemical library information}
\usage{
tcplLoadClib(field = NULL, val = NULL)
}
\arguments{
\item{field}{Character of length 1, \code{'chid'} or \code{'clib'}, whether 
to search by chemical id (chid), or chemical library (clib)}

\item{val}{The values to query on}
}
\value{
A data.table with the chemical library information for the given
parameters.
}
\description{
\code{tcplLoadClib} queries the tcpl databases and returns information 
about the chemial library.
}
\details{
Chemicals are stored in different libraries by chemcial ID. Therefore, it 
is not possible to delineate samples with the same chemical ID into two 
distinct chemical libraries. However, it is possible for a chemcial ID to 
belong to more than one (or no) chemical libraries. 

When chemicals belong to more than one library, the chemical is listed 
multiple times (one for each distinct library).
}
\examples{
## Store the current config settings, so they can be reloaded at the end 
## of the examples
conf_store <- tcplConfList()
tcplConfDefault()

## Passing no parameters gives all of the chemical ISs that have a chemical
## library registered
clib <- tcplLoadClib()

## Notice there are more rows in tcplLoadClib than in tcplLoadChem, 
## indicating some chemicals must belong to more than library.
chem <- tcplLoadChem(include.spid = FALSE)
nrow(chem)
nrow(clib)

## It is possible that some chemicals do not have a chemical library 
## registered, although this is not the case in the example data.
all(chem$chid \%in\% clib$chid)

## Show the unique chemical libraries
clib[ , unique(clib)]

## Specifying a chemical library will not show what other libraries a 
## chemical might belong to.
tcplLoadClib(field = "clib", val = "other")
tcplLoadClib(field = "chid", val = 1:2)

## Reset configuration
options(conf_store)

}
