% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplLoadConcUnit.R
\name{tcplLoadConcUnit}
\alias{tcplLoadConcUnit}
\title{Load concentration units for assay endpoints}
\usage{
tcplLoadConcUnit(spid)
}
\arguments{
\item{spid}{Integer, assay endpoint ids}
}
\value{
A data.table containing level 3 correction methods for the given
spids.
}
\description{
\code{tcplLoadUnit} queries the tcpl databases and returns a data.table
with the concentration units for the given assay endpoint ids (spid).
}
\seealso{
\code{\link{tcplQuery}}, \code{\link{data.table}}
}
