% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplLoadUnit.R
\name{tcplLoadUnit}
\alias{tcplLoadUnit}
\title{Load response units for assay endpoints}
\usage{
tcplLoadUnit(aeid)
}
\arguments{
\item{aeid}{Integer, assay endpoint ids}
}
\value{
A data.table containing level 3 correction methods for the given
aeids.
}
\description{
\code{tcplLoadUnit} queries the tcpl databases and returns a data.table
with the response units for the given assay endpoint ids (aeid).
}
\seealso{
\code{\link{tcplQuery}}, \code{\link{data.table}}
}

