% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotUtils.R
\name{tcplPlotCalcAspectRatio}
\alias{tcplPlotCalcAspectRatio}
\title{tcplPlotCalcAspectRatio}
\usage{
tcplPlotCalcAspectRatio(
  type = "mc",
  verbose = FALSE,
  multi = FALSE,
  nrows = 0,
  output = c("ggplot", "console", "pdf", "png", "jpg", "svg", "tiff"),
  group.threshold = 9,
  nrow = NULL,
  ncol = NULL,
  flags = FALSE
)
}
\arguments{
\item{type}{string of mc or sc indicating if it is single or multi conc}

\item{verbose}{should the plot return a table with parameters}

\item{multi}{Boolean, by default TRUE for "pdf". Prints variable number of plots
per page depending on 'verbose' and 'type' settings.}

\item{nrows}{Integer, number of rows each compare plot uses}

\item{output}{How should the plot be presented. To work with the plot in 
environment, use "ggplot"; to interact with the plot in application, use 
"console"; or to save as a file type, use "pdf", "jpg", "png", "svg", or "tiff".}

\item{group.threshold}{Integer of length 1, minimum number of samples in a 
given plot where comparison plots, instead of coloring models by sample, should
delineate curve color by a given group.fld. By default 9.}

\item{nrow}{Integer, number of rows in multiplot. By default 2.}

\item{ncol}{Integer, number of columns in multiplot. By default 3, 2 if verbose, 
1 for compare plots.}

\item{flags}{Boolean, by default FALSE. If TRUE, level 6 flags are displayed
within output.}
}
\value{
a list of validated parameters for plotting
}
\description{
tcplPlotCalcAspectRatio
}
