% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotUtils.R
\name{tcplPlotLoadWllt}
\alias{tcplPlotLoadWllt}
\title{tcplPlotLoadWllt
Replaces NA dtxsid and chnm with a string description of the sample's well type(s)}
\usage{
tcplPlotLoadWllt(dat = NULL, type = "mc")
}
\arguments{
\item{dat}{dataset}

\item{type}{mc or sc}
}
\value{
dat with updated dtxsid/chnm if they are NA
}
\description{
tcplPlotLoadWllt
Replaces NA dtxsid and chnm with a string description of the sample's well type(s)
}
