% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlot.R
\name{tcplPlotlyPlot}
\alias{tcplPlotlyPlot}
\title{tcplPlotlyPlot}
\usage{
tcplPlotlyPlot(dat, lvl = 5)
}
\arguments{
\item{dat}{data table with all required conc/resp data}

\item{lvl}{integer level of data that should be plotted
level 4 - all fit models
level 5 - all fit models and winning model with hitcall
level 6 - include all flags}
}
\value{
A plotly plot
}
\description{
tcplPlotlyPlot
}
