% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotUtils.R
\name{tcplPlotSetYRange}
\alias{tcplPlotSetYRange}
\title{tcplPlotSetYRange}
\usage{
tcplPlotSetYRange(dat, yuniform, yrange, type)
}
\arguments{
\item{dat}{dataset}

\item{yuniform}{should the yrange be uniform}

\item{yrange}{length 2 of the yrange}

\item{type}{mc or sc}
}
\value{
yrange of the data
}
\description{
tcplPlotSetYRange
}
