% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotUtils.R
\name{tcplPlotValidate}
\alias{tcplPlotValidate}
\title{tcplPlotValidate}
\usage{
tcplPlotValidate(
  type = "mc",
  flags = NULL,
  output = "none",
  multi = NULL,
  verbose = FALSE
)
}
\arguments{
\item{type}{string of mc or sc indicating if it is single or multi conc}

\item{flags}{bool - should we return flags}

\item{output}{how should the plot be formatted}

\item{multi}{are there multiple plots}

\item{verbose}{should the plot return a table with parameters}
}
\value{
a list of validated parameters for plotting
}
\description{
tcplPlotValidate
}
