% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplspidAPIsearch.R
\name{tcplspidAPIsearch}
\alias{tcplspidAPIsearch}
\title{A function for hitting the chemtrack api to get sample details}
\usage{
tcplspidAPIsearch(spid, verbose = FALSE)
}
\arguments{
\item{spid}{single spid or an array of spids}

\item{verbose}{if true we want to return all values from api, not just those that match the sample table format}
}
\value{
a data_table containing the results of hitting the chemtrack api for each spid
}
\description{
A function for hitting the chemtrack api to get sample details
}
\examples{
tcpl:::tcplspidAPIsearch("PINTO01C04")
}
