% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_api}
\alias{test_api}
\title{List containing ids used for different automated tests of tcpl integration
with the CTX APIs, randomly selected from what is available via API.}
\format{
A list with 7 items:
\describe{
  \item{aeid}{Randomly selected assay component endpoint id}
  \item{acid}{Assay component id associated with the above aeid}
  \item{aid}{Assay id associated with the above aeid}
  \item{asid}{Assay source id associated with the above aeid}
  \item{dtxsid}{dsstox substance id of one sample from the above aeid}
  \item{spid}{Sample id of one (the same) sample from the above aeid}
  \item{m4id}{Level 4 id of one (the same) sample from the above aeid}
}
}
\source{
CTX Bioactivity API
}
\usage{
test_api
}
\description{
List containing ids used for different automated tests of tcpl integration
with the CTX APIs, randomly selected from what is available via API.
}
\keyword{datasets}
