% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_schema_functions.R
\name{write_lvl_4}
\alias{write_lvl_4}
\title{Write level 4 with updated schema}
\usage{
write_lvl_4(dat)
}
\arguments{
\item{dat}{output of tcplfit2 that has been unnested into a data.table}
}
\description{
Write level 4 with updated schema
}
