% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chdat}
\alias{chdat}
\title{Chemical library of tested chemicals in the example datasets with the corresponding sample IDs.}
\format{
A data frame with 6 rows and 6 variables:
\describe{
 \item{spid}{sample ID}
  \item{casn}{Chemical Abstract Service(CAS) number}
   \item{chnm}{chemical name}
    \item{dsstox_substance_id}{chemical-specific DTXSID}
     \item{code}{CAS number compressed into numeric string}
      \item{chid}{unique chemical ID number for tcpl}

}
}
\source{
Toxcast database
}
\usage{
chdat
}
\description{
Chemical library of tested chemicals in the example datasets with the corresponding sample IDs.
}
\keyword{datasets}
