% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_schema_functions.R
\name{check_tcpl_db_schema}
\alias{check_tcpl_db_schema}
\title{Function that checks if the most recent v3 table schema is used in the database schema}
\usage{
check_tcpl_db_schema()
}
\value{
boolean TRUE if param tables are listed in schema FALSE otherwise
}
\description{
Function that checks if the most recent v3 table schema is used in the database schema
}
