% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu.R
\name{lu}
\alias{lu}
\title{Abbreviation for \code{length(unique(x))}}
\usage{
lu(x)
}
\arguments{
\item{x}{A logical}
}
\value{
The unique of the \code{TRUE} values in \code{x}
}
\description{
\code{lu} takes a logical vector, \code{x}, and returns 
\code{length(unique(x))}.
}
\seealso{
\code{\link{unique}}, \code{\link{which}}

Other tcpl abbreviations: \code{\link{is.odd}},
  \code{\link{lw}}, \code{\link{sink.reset}}
}
