% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_vignette}
\alias{sc_vignette}
\title{List with single-concentration data for the vignette}
\format{
\enumerate{
\item \strong{sc0} A data frame with 10 rows and 18 columns
\describe{
  \item{spid}{sample ID}
  \item{chid}{unique chemical ID number for tcpl}
  \item{casn}{Chemical Abstract Service(CAS) number}
  \item{chnm}{chemical name}
  \item{dsstox_substance_id}{chemical-specific DTXSID}
  \item{code}{CAS number compressed into numeric string}
  \item{acid}{Assay Component ID}
  \item{acnm}{Assay Component Name}
  \item{s0id}{SC0 ID}
  \item{apid}{assay plate ID}
  \item{rowi}{Row Index}
  \item{coli}{Column Index}
  \item{wllt}{Well Type}
  \item{wllq}{Well Quality (0 or 1)}
  \item{conc}{concentration in micromolar}
  \item{rval}{raw assay component readout value}
  \item{srcf}{source file containing the raw data}
  \item{conc_unit}{Concentration Units}
  }


\item \strong{sc1} A data frame with 10 rows and 20 columns
\describe{
  \item{spid}{sample ID}
  \item{chid}{unique chemical ID number for tcpl}
  \item{casn}{Chemical Abstract Service(CAS) number}
  \item{chnm}{chemical name}
  \item{dsstox_substance_id}{chemical-specific DTXSID}
  \item{code}{CAS number compressed into numeric string}
  \item{aeid}{Assay Component Endpoint ID}
  \item{aenm}{assay endpoint name (i.e., assay_component_endpoint_name)}
  \item{acid}{Assay Component ID}
  \item{acnm}{Assay Component Name}
  \item{s0id}{SC0 ID}
  \item{s1id}{SC1 ID}
  \item{apid}{assay plate ID}
  \item{rowi}{Row Index}
  \item{coli}{Column Index}
  \item{wllt}{Well Type}
  \item{logc}{Log base 10 concentration}
  \item{resp}{Normalized response value}
  \item{resp_unit}{Response Units}
  \item{conc_unit}{Concentration Units}
  }


\item \strong{sc2} A data frame with 10 rows and 15 columns
\describe{
  \item{spid}{sample ID}
  \item{chid}{unique chemical ID number for tcpl}
  \item{casn}{Chemical Abstract Service(CAS) number}
  \item{chnm}{chemical name}
  \item{dsstox_substance_id}{chemical-specific DTXSID}
  \item{code}{CAS number compressed into numeric string}
  \item{aeid}{Assay Component Endpoint ID}
  \item{aenm}{assay endpoint name (i.e., assay_component_endpoint_name)}
  \item{s2id}{SC2 ID}
  \item{bmad}{The median absolute deviation of all treatment wells (default option) or blank wells }
  \item{max_med}{maximum median response}
  \item{hitc}{Hitcall}
  \item{coff}{Cutoff}
  \item{resp_unit}{Response Units}
  \item{conc_unit}{Concentration Units}
  }
}
}
\usage{
sc_vignette
}
\description{
This dataset is a list with 3 data.tables (sc0,sc1,sc2).
}
\keyword{datasets}
