% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplFit2.R
\name{tcplFit2}
\alias{tcplFit2}
\title{tcpl Wrapper for tcplfit2_core including additional calculations to fit into new schema}
\usage{
tcplFit2(
  dat,
  fitmodels = c("cnst", "hill", "gnls", "poly1", "poly2", "pow", "exp2", "exp3",
    "exp4", "exp5"),
  bmed = NULL
)
}
\arguments{
\item{dat}{}
}
\value{
Data.table with an additional column fitparams that includes all of the fitting parameters
}
\description{
tcpl Wrapper for tcplfit2_core including additional calculations to fit into new schema
}
