% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlot.R
\name{tcplggplot}
\alias{tcplggplot}
\title{tcplggplot}
\usage{
tcplggplot(dat, lvl = 5, verbose = FALSE, flags = FALSE)
}
\arguments{
\item{dat}{data table with all required conc/resp data}

\item{lvl}{integer level of data that should be plotted
level 2 - for 'sc' plotting
level 5 - for 'mc' plotting, all fit models and winning model with hitcall}

\item{verbose}{boolean should plotting include table of values next to the plot}

\item{flags}{boolean should plotting include level 6 flags in plot caption}
}
\value{
A ggplot object or grob with accompanied table depending on verbose option
}
\description{
tcplggplot
}
