% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlot.R
\name{tcplPlot}
\alias{tcplPlot}
\title{#-------------------------------------------------------------------------------
 Generic Plotting Function for tcpl}
\usage{
tcplPlot(
  dat = NULL,
  type = "mc",
  fld = "m4id",
  val = NULL,
  compare = "m4id",
  by = NULL,
  output = c("ggplot", "console", "pdf", "png", "jpg", "svg", "tiff"),
  fileprefix = paste0("tcplPlot_", Sys.Date()),
  multi = NULL,
  verbose = TRUE,
  nrow = NULL,
  ncol = NULL,
  dpi = 600,
  flags = FALSE,
  yuniform = FALSE,
  yrange = c(NA, NA),
  group.fld = NULL,
  group.threshold = 9
)
}
\arguments{
\item{dat}{data.table or list of data.tables containing plot-prepared data, not
required. Used for stand-alone (ToxCast or other tcplfit2-fit data) plotting or 
advanced plotting (generating comparison plots across multiple database configurations). 
Pass a data.table for default behavior, which will split data by 'compare'. Pass 
a list of data.tables to directly specify the groupings for comparison plots 
where each list item (data.table) will be printed on a single plot. See \code{tcplPlotLoadData}.}

\item{type}{Character of length 1, the data type, "sc" or "mc".}

\item{fld}{Character vector, the field(s) to query on.}

\item{val}{List containing vectors of values for each field to query on. Must 
be in the same order as 'fld'.}

\item{compare}{Character vector, the field(s) to join samples on to create comparison
plots. By default "m4id", "s2id" if 'type' = "sc". As every endpoint-sample will 
always have its own m4id, this will create individual plots. To create a comparison 
plot across the same chemicals, use a chemical identifier like "dsstox_substance_id". 
Likewise, to create a comparison plot across the same sample ids, use "spid". 
Use "aeid" to create a comparison plot across the same assay component endpoints, 
which will likely trigger the large compare plot style; for more info see 'group.threshold'. 
To use a custom field to create comparison, 'dat' should be supplied as a data.table 
generated from tcplPlotLoadData with the custom column added. If 'dat' is instead 
a list of data.tables, setting 'compare' will be ignored in favor of the list 
groups.}

\item{by}{Parameter to divide files into e.g. "aeid".}

\item{output}{How should the plot be presented. To work with the plot in 
environment, use "ggplot"; to interact with the plot in application, use 
"console"; or to save as a file type, use "pdf", "jpg", "png", "svg", or "tiff".}

\item{fileprefix}{Prefix of file when saving.}

\item{multi}{Boolean, by default TRUE for "pdf". Prints variable number of plots
per page depending on 'verbose' and 'type' settings.}

\item{verbose}{Boolean, by default TRUE. If TRUE, a table with fitting parameters
is included with the plot. To return or save simple plot, set to FALSE. In comparison
plotting, verbose as TRUE will also return annotation information in a table,
which is hidden using verbose = FALSE.}

\item{nrow}{Integer, number of rows in multiplot. By default 2.}

\item{ncol}{Integer, number of columns in multiplot. By default 3, 2 if verbose, 
1 for verbose compare plots.}

\item{dpi}{Integer, image print resolution. By default 600.}

\item{flags}{Boolean, by default FALSE. If TRUE, level 6 flags are displayed
within output.}

\item{yuniform}{Boolean, by default FALSE. If TRUE, all plots will have uniform
y axis scaling, automatically determined.}

\item{yrange}{Integer of length 2, for directly setting the y-axis range, 
c(<min>,<max>). By default, c(NA,NA). Plots containing points or curves outside
this range will be expanded to fit.}

\item{group.fld}{string column name to group curves by when number of comparison 
curves exceeds group.threshold, default being 'modl' for MC and 'hitc' for SC}

\item{group.threshold}{integer of length 1, number of curves where comparison 
plot style should change to instead group by a given group.fld, default of 9 
-- greater than 8 curves}
}
\description{
\code{tcplPlot} queries the tcpl databases and returns a plot
for the given level and data type.
}
\details{
The data type can be either 'mc' for multiple concentration data, or 'sc'
for single concentration data. 

Leaving \code{fld} NULL will return all data.
}
\examples{
\dontrun{
tcplPlot(fld = "m4id", val = c(18609966)) ## Create a level 4 plot
}
}
